🔒️ 치명적인 보안이슈

+++ Anything_back/src/main/java/com/gsmhrm/anything_back/domain/plan/service/DeletePlanService.java
@@ -1,9 +1,11 @@
 package com.gsmhrm.anything_back.domain.plan.service;
 
 import com.gsmhrm.anything_back.domain.plan.entity.Plan;
+import com.gsmhrm.anything_back.domain.plan.exception.NoDeletePermissionException;
 import com.gsmhrm.anything_back.domain.plan.exception.NotFoundPlanException;
 import com.gsmhrm.anything_back.domain.plan.repository.PlanRepository;
 import com.gsmhrm.anything_back.global.annotation.RollbackService;
+import com.gsmhrm.anything_back.global.util.UserUtil;
 import lombok.RequiredArgsConstructor;
 
 @RollbackService
@@ -11,11 +13,16 @@ import lombok.RequiredArgsConstructor;
 public class DeletePlanService {
 
     private final PlanRepository planRepository;
+    private final UserUtil util;
 
     public void execute(Long id) {
 
         Plan plan = planRepository.findById(id)
                 .orElseThrow(NotFoundPlanException::new);
+
+        if (!(plan.getMember() == util.currentUser())) {
+            throw new NoDeletePermissionException();
+        }
 
         planRepository.deleteById(id);
     }
